// src/models/student.model.ts
export interface IStudent {
    key: string;
    name: string; //student name
    fname: string; // father name
    phNumber: string; // phone number
    currentClass: string; // class of student
    
   date1 : string; // fees for month of Jan
   date2 : string; // for feb
   date3 : string; // for march and so on...
} 